+++ smos/src/Smos/Default.hs
@@ -135,6 +135,7 @@ defaultFileKeyMap =
             exactString "zh" forestToggleCollapseEntryHistory,
             exactString "zl" forestToggleCollapseEntryLogbook,
             exactString "zt" forestToggleCollapseEntryTimestamps,
+            exactString "zp" forestToggleCollapseEntryProperties,
             -- Entering contents
             exactKey KEnter entrySelectContentsAtEnd,
             exactKeyPress (KeyPress KEnter [MMeta]) entrySelectContentsAtStart,